backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.6.0",
    "appnope==0.1.4",
    "argon2-cffi==23.1.0",
    "argon2-cffi-bindings==21.2.0",
    "asttokens==2.4.1",
    "attrs==24.2.0",
    "backcall==0.2.0",
    "beautifulsoup4==4.12.3",
    "bleach==6.2.0",
    "certifi==2024.8.30",
    "cffi==1.17.1",
    "charset-normalizer==3.4.0",
    "click==8.1.7",
    "cryptography==43.0.1",
    "decorator==5.1.1",
    "defusedxml==0.7.1",
    "dnspython==2.7.0",
    "docopt==0.6.2",
    "email-validator==2.2.0",
    "executing==2.1.0",
    "fastapi==0.115.2",
    "fastapi-cli==0.0.5",
    "fastjsonschema==2.20.0",
    "h11==0.14.0",
    "httpcore==1.0.6",
    "httptools==0.6.1",
    "httpx==0.27.2",
    "idna==3.10",
    "ipython==8.12.3",
    "jedi==0.19.2",
    "jinja2==3.1.4",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "jupyterlab-pygments==0.3.0",
    "loguru==0.7.2",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.1",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mistune==3.0.2",
    "mypy==1.13.0",
    "mypy-extensions==1.0.0",
    "nbclient==0.10.0",
    "nbconvert==7.16.4",
    "nbformat==5.10.4",
    "packaging==24.2",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pickleshare==0.7.5",
    "pipreqs==0.5.0",
    "platformdirs==4.3.6",
    "prompt-toolkit==3.0.48",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pycparser==2.22",
    "pydantic==2.9.2",
    "pydantic-core==2.23.4",
    "pygments==2.18.0",
    "pyjwt==2.9.0",
    "pymysql==1.1.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.12",
    "pyyaml==6.0.2",
    "pyzmq==26.2.0",
    "referencing==0.35.1",
    "requests==2.32.3",
    "rich==13.9.2",
    "rpds-py==0.21.0",
    "shellingham==1.5.4",
    "six==1.16.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "sqlalchemy==2.0.35",
    "stack-data==0.6.3",
    "starlette==0.39.2",
    "tinycss2==1.4.0",
    "tornado==6.4.2",
    "traitlets==5.14.3",
    "typer==0.12.5",
    "typing-extensions==4.12.2",
    "urllib3==2.2.3",
    "uvicorn==0.31.1",
    "uvloop==0.20.0",
    "watchfiles==0.24.0",
    "wcwidth==0.2.13",
    "webencodings==0.5.1",
    "websockets==13.1",
    "yarg==0.1.9",
]
